Minor fixes

+++ course-2021-1/exercises/04-wubba-lubba-dub-dub/WubbaLubbaDubDub/WubbaLubbaDubDub/RicksMercilessEncryptor.cs
@@ -112,7 +112,7 @@ namespace WubbaLubbaDubDub
         {
             char Shift(char c)
             {
-                return (char) ((int) c + 1);
+                return (char) (c + 1);
             }
 
             return String.Concat(s.Select(Shift));
